backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.15.0,<3.16.0
django-cors-headers>=4.0.0,<5.0.0
django-environ>=0.11.0,<1.0.0
mysqlclient>=2.0.0,<3.0.0
# chromadb>=0.5.0,<0.6.0
# sentence-transformers>=3.0.0,<4.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.0.0,<1.0.0
selenium>=4.0.0,<5.0.0
webdriver-manager>=4.0.0,<5.0.0
beautifulsoup4>=4.0.0,<5.0.0
celery>=5.0.0,<6.0.0
redis>=5.0.0,<6.0.0
django-redis>=5.0.0,<6.0.0
Pillow>=10.0.0,<11.0.0
requests>=2.0.0,<3.0.0
python-decouple>=3.0.0,<4.0.0
